implementing geolocator

+++ lib/data/position/geolocator/model/position_geolocator_model.dart
@@ -1,20 +1,11 @@
 class PositionGeolocatorModel {
-  final int longitude;
-  final int latitude;
-  final int accuracy;
+  final double longitude;
+  final double latitude;
+  final double accuracy;
 
   PositionGeolocatorModel({
     required this.longitude,
     required this.latitude,
     required this.accuracy,
   });
-
-  // PositionGeolocatorModel.fromJson(Map<String, dynamic> json)
-  //     : singleFlips = json['singleFlips'] as int,
-  //       solutionsNumber = json['solutionsNumber'] as int;
-
-  // Map<String, dynamic> toJson() => {
-  //       'singleFlips': singleFlips,
-  //       'solutionsNumber': solutionsNumber,
-  //     };
 }
